+++ src/main/java/dev/artemcodes/store_ecom/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package dev.artemcodes.store_ecom.repositories;
 import dev.artemcodes.store_ecom.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Byte> {
+public interface CategoryRepository extends CrudRepository<Category, Short> {
 }
